{{cookiecutter.project_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_verbose_name}}"
version = "0.1.0"
description = ""
authors = ["Waseem akram <waseem07799@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
aiohttp = {extras = ["speedups"], version = "^3.8.3"}
argon2-cffi = "^21.3.0"
celery = "^5.2.7"
django = "^4.1.8"
django-admin-interface = "^0.22.1"
django-allauth = "^0.54.0"
django-anymail = {extras = ["amazon-ses"], version = "^8.6"}
django-browser-reload = "^1.7.0"
django-celery-email="^3.0.0"
django-csp = "^3.7"
django-environ = "^0.9.0"
django-feather = "^0.3.0"
django-htmx = "^1.12.2"
django-http-referrer-policy = "^1.1.1"
django-mjml = "^1.0"
django-model-utils = "^4.2.0"
django-nested-admin = "^4.0.2"
django-permissions-policy = "^4.13.0"
django-phonenumber-field = { extras = ["phonenumberslite"], version = "^7.0.0" }
django-redis = "^5.2.0"
django-render-block = "^0.9.2"
django-split-settings = "^1.2.0"
django-storages = {extras = ["boto3"], version = "^1.13.1"}
django-structlog = "^4.0.1"
django-vite = "^2.0.2"
django-widget-tweaks = "^1.4.12"
easy-thumbnails = "^2.8.3"
geojson = "^2.5.0"
gunicorn = "^20.1.0"
hiredis = "^2.0.0"
numerize = "^0.12"
psycopg2 = "^2.9.5"
pydantic = "^1.10.2"
python-dateutil = "^2.8.2"
redis = "^4.3.4"
structlog = "^22.1.0"
types-python-dateutil = "^2.8.19.3"
uvicorn = { extras = ["standard"], version = "^0.19.0" }
django-dbbackup = "^4.0.2"
django-ipware = "^5.0.0"
django-ratelimit = "^4.1.0"


[tool.poetry.group.local.dependencies]
argparse = "^1.4.0"
black = "^22.10.0"
django-coverage-plugin = "^3.0.0"
django-cprofile-middleware = "^1.0.5"
django-debug-toolbar = "^3.7.0"
django-extensions = "^3.2.1"
django-silk = "^5.0.2"
django-stubs = {extras = ["compatible-mypy"], version = "^1.14.0"}
djhtml = "^3.0.6"
djlint = "^1.19.13"
fabric2 = "^3.0.0"
flake8 = "^6.0.0"
hypothesis = {extras = ["django"], version = "^6.61.0"}
ipdb = "^0.13.9"
jupyterlab = "^3.5.0"
mixer = "^7.2.2"
neovim = "^0.3.1"
nplusone = "^1.0.0"
pre-commit = "^2.20.0"
pudb = "^2022.1.3"
pylint = "^2.15.9"
pylint-django = "^2.5.3"
pylint-pytest = "^1.1.2"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.10.0"
pytest-sugar = "^0.9.6"
pytest-testmon = "^1.4.2"
pytest-xdist = "^3.1.0"
types-requests = "^2.28.11.15"
watchdog = "^2.2.0"
werkzeug = "^2.2.3"

[tool.pyright]
exclude = [ ".venv" ]
venvPath = "."
venv = ".venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
